test: canary5-v2 = McpDashboardsLib NAME + full dashboards body (combination probe)
canary4 (full body, McpCanary4Lib) bound; canary5 (McpDashboardsLib, minimal) bound. canary5-v2 is
the one untested pairing: the McpDashboardsLib name WITH the full body, marker-probed via hub_get_info
(part-methods inert, no aggregator wiring; canary4 removed to avoid duplicate-method collision). If
mcpDashNameProbe fails to bind -> the name+content combination is the breaker. If it binds -> the
combination is fine and the dashboards failure is the aggregator-registration wiring in the app.
